+++ scripts/release/release.js
@@ -14,10 +14,15 @@ async function run() {
   const { runGit, readJson } = await import("./utils.js");
 
   const params = minimist(process.argv.slice(2), {
-    string: ["version"],
-    boolean: ["dry", "manual"],
+    string: ["version", "repo"],
+    boolean: ["dry", "manual", "skip-dependencies-install"],
     alias: { v: "version" },
-    default: { manual: false },
+    default: {
+      manual: false,
+      dry: false,
+      "skip-dependencies-install": false,
+      repo: "git@github.com:prettier/prettier.git",
+    },
   });
 
   const { stdout: previousVersion } = await runGit([
@@ -39,7 +44,7 @@ async function run() {
     [
       "./steps/validate-new-version.js",
       "./steps/check-git-status.js",
-      "./steps/install-dependencies.js",
+      !params["skip-dependencies-install"] && "./steps/install-dependencies.js",
       params.manual && "./steps/run-tests.js",
       "./steps/update-version.js",
       params.manual && "./steps/generate-bundles.js",