+++ src/cli/format.js
@@ -7,6 +7,9 @@ import { createIgnorer, errors } from "./prettier-internal.js";
 import { expandPatterns, fixWindowsSlashes } from "./expand-patterns.js";
 import getOptionsForFile from "./options/get-options-for-file.js";
 import isTTY from "./is-tty.js";
+import findCacheFile from "./find-cache-file.js";
+import FormatResultsCache from "./format-results-cache.js";
+import { statSafe } from "./utils.js";
 
 const { getStdin } = thirdParty;
 
@@ -301,6 +304,26 @@ async function formatFiles(context) {
     context.logger.log("Checking formatting...");
   }
 
+  let formatResultsCache;
+  const cacheFilePath = findCacheFile();
+  if (context.argv.cache) {
+    formatResultsCache = new FormatResultsCache(
+      cacheFilePath,
+      context.argv.cacheStrategy || "content"
+    );
+  } else {
+    if (context.argv.cacheStrategy) {
+      context.logger.error(
+        "`--cache-strategy` cannot be used without `--cache`."
+      );
+      process.exit(2);
+    }
+    const stat = await statSafe(cacheFilePath);
+    if (stat) {
+      await fs.unlink(cacheFilePath);
+    }
+  }
+
   for await (const pathOrError of expandPatterns(context)) {
     if (typeof pathOrError === "object") {
       context.logger.error(pathOrError.error);
@@ -368,16 +391,31 @@ async function formatFiles(context) {
 
     const start = Date.now();
 
+    const isCacheExists = formatResultsCache?.existsAvailableFormatResultsCache(
+      filename,
+      options
+    );
+
     let result;
     let output;
 
     try {
+<<<<<<< HEAD
       result = await format(context, input, options);
+=======
+      if (isCacheExists) {
+        result = { formatted: input };
+      } else {
+        result = format(context, input, options);
+      }
+>>>>>>> main
       output = result.formatted;
     } catch (error) {
       handleError(context, filename, error, printedFilename);
       continue;
     }
+
+    formatResultsCache?.setFormatResultsCache(filename, options);
 
     const isDifferent = output !== input;
 
@@ -414,7 +452,12 @@ async function formatFiles(context) {
           process.exitCode = 2;
         }
       } else if (!context.argv.check && !context.argv.listDifferent) {
-        context.logger.log(`${chalk.grey(filename)} ${Date.now() - start}ms`);
+        const message = `${chalk.grey(filename)} ${Date.now() - start}ms`;
+        if (isCacheExists) {
+          context.logger.log(`${message} (cached)`);
+        } else {
+          context.logger.log(message);
+        }
       }
     } else if (context.argv.debugCheck) {
       /* istanbul ignore else */
@@ -436,6 +479,8 @@ async function formatFiles(context) {
       numberOfUnformattedFilesFound += 1;
     }
   }
+
+  formatResultsCache?.reconcile();
 
   // Print check summary based on expected exit code
   if (context.argv.check) {