+++ src/cli/utils.js
@@ -2,6 +2,11 @@ import { __internal as sharedWithCli } from "../index.js";
 
 const { arrayify, getLast, isNonEmptyArray, partition } = sharedWithCli.utils;
 
+const { promises: fs } = require("fs");
+
+// eslint-disable-next-line no-restricted-modules
+const { default: sdbm } = require("../../vendors/sdbm.js");
+
 // eslint-disable-next-line no-console
 const printToScreen = console.log.bind(console);
 
@@ -40,6 +45,30 @@ function pick(object, keys) {
   return Object.fromEntries(entries);
 }
 
+/**
+ * @param {string} source
+ * @returns {string}
+ */
+function createHash(source) {
+  return String(sdbm(source));
+}
+
+/**
+ * Get stats of a given path.
+ * @param {string} filePath The path to target file.
+ * @returns {Promise<import('fs').Stats | undefined>} The stats.
+ */
+async function statSafe(filePath) {
+  try {
+    return await fs.stat(filePath);
+  } catch (error) {
+    /* istanbul ignore next */
+    if (error.code !== "ENOENT") {
+      throw error;
+    }
+  }
+}
+
 export {
   arrayify,
   getLast,
@@ -48,4 +77,5 @@ export {
   printToScreen,
   groupBy,
   pick,
+createHash, statSafe
 };